Add RawSuggestionPostProcessingStrategy
