+++ Core/Sources/SuggestionService/RawSuggestionPostProcessing/NoOpRawSuggestionPostProcessingStrategy.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct NoOpRawSuggestionPostProcessingStrategy: RawSuggestionPostProcessingStrategy {
+    func postProcessRawSuggestion(suggestionPrefix: String, suggestion: String) -> String {
+        suggestionPrefix + suggestion
+    }
+}
+
